Add commands ih-ec2 launch-templates and subnets
