+++ infrahouse_toolkit/cli/ih_ec2/__init__.py
@@ -21,7 +21,9 @@ from infrahouse_toolkit.aws import aws_sso_login
 from infrahouse_toolkit.aws.config import AWSConfig
 from infrahouse_toolkit.cli.ih_ec2.cmd_instance_types import cmd_instance_types
 from infrahouse_toolkit.cli.ih_ec2.cmd_launch import cmd_launch
+from infrahouse_toolkit.cli.ih_ec2.cmd_launch_templates import cmd_launch_templates
 from infrahouse_toolkit.cli.ih_ec2.cmd_list import cmd_list
+from infrahouse_toolkit.cli.ih_ec2.cmd_subnets import cmd_subnets
 from infrahouse_toolkit.cli.ih_ec2.cmd_terminate import cmd_terminate
 from infrahouse_toolkit.logging import setup_logging
 
@@ -115,6 +117,7 @@ def ih_ec2(ctx, **kwargs):
         ctx.obj = {
             "debug": kwargs["debug"],
             "ec2_client": get_aws_client("ec2", aws_profile, aws_region, session=aws_session),
+            "aws_config": aws_config,
         }
     except NoRegionError as err:
         LOG.error(err)
@@ -122,7 +125,7 @@ def ih_ec2(ctx, **kwargs):
         sys.exit(1)
 
 
-for cmd in [cmd_launch, cmd_list, cmd_instance_types, cmd_terminate]:
+for cmd in [cmd_launch, cmd_list, cmd_instance_types, cmd_terminate, cmd_subnets, cmd_launch_templates]:
     # noinspection PyTypeChecker
     ih_ec2.add_command(cmd)
